requirements.txt
flask>=2.0.1
streamlit>=1.28.0
langchain-groq>=0.0.3
python-dotenv==1.0.0
SpeechRecognition>=3.10.0
Pillow>=10.0.0
clarifai-grpc>=9.10.0
scipy>=1.11.0
numpy>=1.24.0
sounddevice>=0.4.6
requests>=2.31.0
protobuf>=4.25.0
grpcio>=1.59.0
langchain>=0.1.0
pyttsx3>=2.90
wave>=0.0.2 